# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Campaign API",
        "description": "API documentation for Campaign management system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/crawl-jobs": {
            "get": {
                "tags": [
                    "Crawl Jobs"
                ],
                "summary": "Get paginated crawl jobs",
                "operationId": "a0def5a9d4627c07dafd9d5aa263f394",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Crawl Jobs"
                ],
                "summary": "Create crawl job",
                "operationId": "730be85cf042be25772ba3175a13fecf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CrawlJob"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/crawl-jobs/{id}": {
            "get": {
                "tags": [
                    "Crawl Jobs"
                ],
                "summary": "Get crawl job by ID",
                "operationId": "8d06f80bc3698580a4c7d3d78a082121",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Crawl Jobs"
                ],
                "summary": "Update crawl job",
                "operationId": "9c36aaed93bc3362fd7a398d3d10bd09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CrawlJob"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Crawl Jobs"
                ],
                "summary": "Delete crawl job",
                "operationId": "2d3062fc443188eba107c55bc45ba0bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/crawl-jobs/status/{status}": {
            "get": {
                "tags": [
                    "Crawl Jobs"
                ],
                "summary": "Get crawl jobs by status",
                "operationId": "1ac25ba3ebc8430ab9254e5e20a3b543",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "in_progress",
                                "success",
                                "failed"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/crawl-jobs/campaign/{campaignId}": {
            "get": {
                "tags": [
                    "Crawl Jobs"
                ],
                "summary": "Get crawl jobs by campaign",
                "operationId": "0b0a234d83889ed29a9389141f13a90e",
                "parameters": [
                    {
                        "name": "campaignId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/crawl-jobs/source/{sourceId}": {
            "get": {
                "tags": [
                    "Crawl Jobs"
                ],
                "summary": "Get crawl jobs by source",
                "operationId": "67cfd5ac5ad89cb1169251821c5eb958",
                "parameters": [
                    {
                        "name": "sourceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CrawlJob": {
                "description": "Crawl job model",
                "properties": {
                    "id": {
                        "description": "Crawl job ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "campaign_id": {
                        "description": "Associated campaign ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1,
                        "nullable": true
                    },
                    "source_id": {
                        "description": "Associated source ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1,
                        "nullable": true
                    },
                    "started_at": {
                        "description": "Job start timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "finished_at": {
                        "description": "Job finish timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T01:00:00Z",
                        "nullable": true
                    },
                    "status": {
                        "description": "Job status",
                        "type": "string",
                        "enum": [
                            "pending",
                            "in_progress",
                            "success",
                            "failed"
                        ],
                        "example": "pending"
                    },
                    "total_articles": {
                        "description": "Total articles crawled",
                        "type": "integer",
                        "example": 0
                    },
                    "error_message": {
                        "description": "Error message if failed",
                        "type": "string",
                        "example": "Connection timeout",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}